requirements.txt
sentence-transformers
chromadb
transformers
torch
fastapi
uvicorn
pymongo
python-dotenv
sentencepiece
